Clicking on an <input type="image"> will submit the form with null submitter https://bugs.webkit.org/show_bug.cgi?id=236324
LayoutTests/imported/w3c:

Add a test to make sure that when the form is submitted through an <input type="image">
control, the `submitter` field of the submit event is that control.

Patch by Andreu Botella <andreu@andreubotella.com> on 2022-02-11
Reviewed by Carlos Garcia Campos.

* web-platform-tests/html/semantics/forms/form-submission-0/form-submission-algorithm-expected.txt:
* web-platform-tests/html/semantics/forms/form-submission-0/form-submission-algorithm.html:

Source/WebCore:

Patch by Andreu Botella <andreu@andreubotella.com> on 2022-02-11
Reviewed by Carlos Garcia Campos.

If you submit a form by clicking on an <input type="image"> control, the control will be in
the form's entry list, but the submitter field in the submit event will be null. This is
wrong per the spec, and caused because the call to `HTMLFormElement::submitIfPossible()` in
`ImageInputType::handleDOMActivateEvent()` calls it with one argument, rather than passing
the element as the submitter.

Tests: imported/w3c/web-platform-tests/html/semantics/forms/form-submission-0/form-submission-algorithm.html

* html/ImageInputType.cpp:
(WebCore::ImageInputType::handleDOMActivateEvent):

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@289615 268f45cc-cd09-0410-ab3c-d52691b4dbfc