[libpas] compact pointers should load payload via memcpy https://bugs.webkit.org/show_bug.cgi?id=236621
Reviewed by Mark Lam.

ASan found that we are loading 8 bytes which can potentially be overflowed.
For example, we load this from a pointer to stack variable, which means that
we could cause stack overflow. Instead we should use memcpy.
Currently, we only support little endian code, but it is OK since libpas is not
enabled in non little endian architectures.

* libpas/src/libpas/pas_compact_ptr.h:
* libpas/src/libpas/pas_compact_tagged_ptr.h:

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@289788 268f45cc-cd09-0410-ab3c-d52691b4dbfc