[Flatpak SDK] Include sparkle-cdm library https://bugs.webkit.org/show_bug.cgi?id=236562
Patch by Philippe Normand <pnormand@igalia.com> on 2022-02-15
Reviewed by Martin Robinson.

This new recipe ships the Sparkle-CDM library, providing an alternative to Thunder. A
ClearKey plugin is provided as well, it will be used later on by the Thunder decryptor,
instead of the dedicated decryptor maintained in WebCore. The goals are to reduce GStreamer
decryptors maintenance here and to dog-food the Thunder code paths using the clearkey layout
tests.

* elements/sdk-platform.bst:
* elements/sdk/sparkle-cdm.bst: Added.

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@289798 268f45cc-cd09-0410-ab3c-d52691b4dbfc