[Forms] Make sure the element's dirty checkedness flag is set to true when setting checked https://bugs.webkit.org/show_bug.cgi?id=236497
Reviewed by Chris Dumez.

LayoutTests/imported/w3c:

Update the test expectations as the tests are now passing.
* web-platform-tests/html/semantics/forms/the-input-element/cloning-steps-expected.txt:

Source/WebCore:

As per spec
https://html.spec.whatwg.org/multipage/input.html#the-input-element%3Aconcept-node-clone-ext,
"The checked IDL attribute allows scripts to manipulate the checkedness of an input element.
... on setting, it must set the element's checkedness to the new value and set the element's
dirty checkedness flag to true." This CL is to make sure this flag is set to true.

* html/HTMLInputElement.cpp:
(WebCore::HTMLInputElement::setChecked):


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@289812 268f45cc-cd09-0410-ab3c-d52691b4dbfc