Defer TerminationsExceptions while in operationMaterializeObjectInOSR. https://bugs.webkit.org/show_bug.cgi?id=236686 rdar://81337114
Reviewed by Saam Barati.

JSTests:

These tests are identical except that they are customized with different watchdog
timeout periods for a Debug / Release build.  This is a necessary condition in
order for the test to manifest this issue if the code is regressed.

* stress/termination-exception-in-operationMaterializeObjectInOSR-debug.js: Added.
* stress/termination-exception-in-operationMaterializeObjectInOSR-release.js: Added.

Source/JavaScriptCore:

operationMaterializeObjectInOSR expects to always succeed.  It is difficult (and
not worth the effort) to make it be able to handle interruptions by the
TerminationException.  Since operationMaterializeObjectInOSR is guaranteed to
finish running in some finite time, it is reasonable to just defer handling a
pending TerminationException until the function returns.

* ftl/FTLOperations.cpp:
(JSC::FTL::JSC_DEFINE_JIT_OPERATION):



git-svn-id: http://svn.webkit.org/repository/webkit/trunk@289877 268f45cc-cd09-0410-ab3c-d52691b4dbfc