Web Inspector: [Flexbox] Gaps and free space do not have pattern fills when navigating to a page with Web Inspector already open https://bugs.webkit.org/show_bug.cgi?id=236741
Reviewed by Devin Rousso.

Because calls to `getDocument` actually resets the `InspectorDOMAgent`, we don't want to clear cached layout
information inside `reset`. The cached layout information is in a WeakHashMap, we aren't going to end up keeping
stale references around anyways, and the under most circumstances the map will clean itself up as it is used.
We should, however go ahead and clear the cached layout information when we are forcing a re-layout to collect
the information again for a new document.

Additionally, if a new document is set, we want to trigger a layout just like we did for a fresh
`InspectorDOMAgent`.

* inspector/agents/InspectorDOMAgent.cpp:
(WebCore::InspectorDOMAgent::didCreateFrontendAndBackend):
(WebCore::InspectorDOMAgent::reset):
(WebCore::InspectorDOMAgent::setDocument):
(WebCore::InspectorDOMAgent::relayoutDocument):
* inspector/agents/InspectorDOMAgent.h:


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@290047 268f45cc-cd09-0410-ab3c-d52691b4dbfc