REGRESSION (r288201): ASSERTION FAILED: !allScriptExecutionContextsMap().contains(m_identifier) https://bugs.webkit.org/show_bug.cgi?id=236123 <rdar://problem/88472494>
Reviewed by Chris Dumez.

We sometimes call setupForReplace which reset m_gotFirstByte to false.
In that case, we may reuse the same client ID for two different documents.
To prevent this, we clear the clientID in setupForReplace.
For good measure, if service worker runtime flag is not on, we do not set the client ID at all.

Covered by existing WK1 tests.

* loader/DocumentLoader.cpp:
(WebCore::DocumentLoader::setupForReplace):
(WebCore::DocumentLoader::loadMainResource):


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@290167 268f45cc-cd09-0410-ab3c-d52691b4dbfc