Set custom general storage directory for custom WebsiteDataStore in TestController https://bugs.webkit.org/show_bug.cgi?id=236893
Reviewed by Chris Dumez.

If we don't set custom generalStorageDirectory for a custom WebsiteDataStore, it will use the same path as the
default WebsiteDataStore. That means when WebKitTestRunner creates multiple WebsiteDataStores (it may create
mutliple TestControllers to run tests in parallel), the directory will be accessed concurrently and data can be
corrupted.

* WebKitTestRunner/TestController.cpp:
(WTR::TestController::configureWebsiteDataStoreTemporaryDirectories):


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@290218 268f45cc-cd09-0410-ab3c-d52691b4dbfc